create cpp file for misc functions
create cpp file for `splitString` and `removeSpaces`
