+++ include/picongpu/plugins/misc/removeSpaces.hpp
@@ -30,22 +30,10 @@ namespace misc
 {
     /** removes all spaces within a string
      *
-     * @param value imput string
+     * @param value input string
      * @return string without any spaces
      */
-    std::string removeSpaces( std::string value )
-    {
-        value.erase(
-            std::remove(
-                value.begin(),
-                value.end(),
-                ' '
-            ),
-            value.end()
-        );
-
-        return value;
-    }
+    std::string removeSpaces( std::string value );
 } // namespace misc
 } // namespace plugins
 } // namespace picongpu
